fix mach

+++ controllers/product.js
@@ -291,9 +291,12 @@ module.exports = {
         try {
             log.info(`Start deleteProduct. Data: ${JSON.stringify(req.body)}`);
             const {_id} = req.body
-
-            await productService.deleteProduct({_id:_id});
-
+            let product = await productService.getProduct({where:{_id:_id}});
+            if(product && product.status === config.PRODUCT_STATUSES.ACTIVE || product && product.status === config.PRODUCT_STATUSES.DRAFT){
+                await  productService.updateProduct({status:config.PRODUCT_STATUSES.ARCHIVED}, {_id:_id});
+            }else{
+                await productService.deleteProduct({_id:_id});
+            }
             log.info(`End deleteProduct`);
 
             return res.status(201).json({delete:_id});
